Data and System Management/requirements.txt
Click==7.0
cycler==0.10.0
dominate==2.3.5
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
paho-mqtt==1.4.0
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
SQLAlchemy==1.3.5
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
